COM-8341: Avoid calling getb2b user api.
We are not using this get b2b account API.
So removed this.
We are calling getUsers API in getB2Baccount api success which  is causing weired issue of sending getB2Baccount reponse as getUsers api request
We need to find different solution to fix such issues.
eg: by reInitilizing model
But in this case we do not need getB2B api  so removing this will fix the issue..